feat: add cookies validation

+++ apps/api/src/controllers/adminAuthorization/delete.Store.Admin.ts
@@ -1,21 +1,26 @@
 import prisma from '@/prisma';
 import { Request, Response } from 'express';
+import { jwtDecode } from 'jwt-decode';
+
+export interface jwtPayload {
+  id: number;
+  role: string;
+}
 
 export const DeleteStoreAdmin = async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
     //cek role dari cookies
     //validasi superadmin
-    // const getCookies = req.cookies.SuperAdminCookie;
-    // const cookiesToDecode = jwtDecode<jwtPayload>(getCookies);
-    // const idSuperAdmin = cookiesToDecode.role;
+    const getCookies = req.cookies['user-token'];
+    const cookiesToDecode = jwtDecode<jwtPayload>(getCookies);
 
-    // if (idSuperAdmin.role !== 'superadmin') {
-    //   return res.status(401).json({
-    //     code: 401,
-    //     message: "you're not authorized",
-    //   });
-    // }
+    if (!cookiesToDecode) {
+      return res.status(401).json({
+        code: 401,
+        message: "you're not authorized",
+      });
+    }
 
     await prisma.storeAdmin.delete({
       where: {
